+++ Boekingssysteem/Boekingssysteem/ApiModels/RoomApiModel.cs
@@ -17,12 +17,12 @@ namespace Boekingssysteem.ApiModels
         public DateTime reservedTill { get; set; }
         public int typeOfRoom { get; set; }
 
-        public RoomApiModel(int hotelId, int roomNumber, int amountOfPeople, decimal price, DateTime reservedFrom, DateTime reservedTill, int typeOfRoom)
+        public RoomApiModel(int hotelId, int roomNumber, int amountOfPeople, double price, DateTime reservedFrom, DateTime reservedTill, int typeOfRoom)
         {
             this.hotelId = hotelId;
             this.roomNumber = roomNumber;
             this.amountOfPeople = amountOfPeople;
-            this.price = price;
+            this.price = (decimal)price;
             this.reservedFrom = reservedFrom;
             this.reservedTill = reservedTill;
             this.typeOfRoom = typeOfRoom;
